Split hal and lib - uefi
