+++ chipsec/lib/uefi_common.py
