+++ chipsec/lib/uefi_compression.py
