+++ chipsec/lib/uefi_fv.py
@@ -26,7 +26,7 @@ import hashlib
 import struct
 from uuid import UUID
 from chipsec.defines import bytestostring
-from chipsec.hal.uefi_common import get_3b_size, bit_set, align
+from chipsec.lib.uefi_common import get_3b_size, bit_set, align
 from chipsec.logger import logger
 
 ################################################################################################
