+++ chipsec/modules/tools/uefi/scan_blocked.py
@@ -52,7 +52,7 @@ from chipsec.module_common import BaseModule, ModuleResult, MTAG_BIOS
 from chipsec.hal.spi_uefi import search_efi_tree, build_efi_model, EFIModuleType
 from chipsec.hal.uefi import UEFI
 from chipsec.hal.spi import SPI, BIOS
-from chipsec.hal.uefi_search import check_match_criteria
+from chipsec.lib.uefi_search import check_match_criteria
 from chipsec.file import read_file, get_main_dir
 
 TAGS = [MTAG_BIOS]
