+++ chipsec/modules/tools/uefi/scan_image.py
@@ -58,7 +58,7 @@ from chipsec.module_common import BaseModule, ModuleResult, MTAG_BIOS
 
 from chipsec.hal.uefi import UEFI
 from chipsec.hal.spi import SPI, BIOS
-from chipsec.hal.uefi_fv import EFI_SECTION, SECTION_NAMES, EFI_SECTION_PE32
+from chipsec.lib.uefi_fv import EFI_SECTION, SECTION_NAMES, EFI_SECTION_PE32
 from chipsec.hal.spi_uefi import build_efi_model, search_efi_tree, EFIModuleType, UUIDEncoder
 from chipsec.file import write_file, read_file
