+++ chipsec/utilcmd/uefi_cmd.py
@@ -58,15 +58,15 @@ import uuid
 from argparse import ArgumentParser
 
 from chipsec.command    import BaseCommand
-from chipsec.hal.uefi_common import EFI_STATUS_DICT, parse_efivar_file
+from chipsec.lib.uefi_common import EFI_STATUS_DICT, parse_efivar_file
 from chipsec.file       import write_file, read_file
 from chipsec.hal.spi_uefi import decode_uefi_region, modify_uefi_region, compress_image, CMD_UEFI_FILE_REPLACE
 from chipsec.hal.spi_uefi import CMD_UEFI_FILE_INSERT_AFTER, CMD_UEFI_FILE_INSERT_BEFORE, CMD_UEFI_FILE_REMOVE
 from chipsec.hal.uefi import UEFI, decode_EFI_variables, get_attr_string, identify_EFI_NVRAM
 from chipsec.hal.uefi import SECURE_BOOT_KEY_VARIABLES, parse_script, parse_EFI_variables
-from chipsec.hal.uefi_fv import get_guid_bin, assemble_uefi_file, assemble_uefi_section, assemble_uefi_raw
-from chipsec.hal.uefi_fv import FILE_TYPE_NAMES
-from chipsec.hal.uefi_platform import fw_types
+from chipsec.lib.uefi_fv import get_guid_bin, assemble_uefi_file, assemble_uefi_section, assemble_uefi_raw
+from chipsec.lib.uefi_fv import FILE_TYPE_NAMES
+from chipsec.lib.uefi_platform import fw_types
 
 
 # Unified Extensible Firmware Interface (UEFI)
